+++ sound-mind-studio/tests/main.cpp
@@ -53,7 +53,6 @@
 #include "test_tool_configuration_panel.h"
 #include "test_tool_palette_controller.h"
 #include "test_undo_stack.h"
-#include "test_warp_dialog.h"
 
 int main(int argc, char** argv) {
     // Redirects ini-format QSettings storage (MainWindow's recentProjects_,
@@ -188,9 +187,6 @@ int main(int argc, char** argv) {
 
     UndoStackTest undoStackTest;
     status |= QTest::qExec(&undoStackTest, argc, argv);
-
-    WarpDialogTest warpDialogTest;
-    status |= QTest::qExec(&warpDialogTest, argc, argv);
 
     AboutDialogTest aboutDialogTest;
     status |= QTest::qExec(&aboutDialogTest, argc, argv);